refactor: ListMenu Dto 수정
